requirements.txt
# Core dependencies
pymongo
faiss-cpu
numpy
sentence-transformers
python-dotenv

# FastAPI dependencies
fastapi
uvicorn
python-multipart
pydantic

# Flask (for backward compatibility)
flask
jinja2

# Utilities
tqdm
loguru
diskcache
